Conflicts:
	cms/development.ini
	cms/src/altaircms/page/api.py
	cms/src/altaircms/page/views.py
	cms/src/altairsite/mobile/__init__.py
	deploy/baseconf/cms-common.ini

+++ cms/src/altaircms/page/views.py
@@ -5,7 +5,8 @@ import sqlalchemy as sa
 import sqlalchemy.orm as orm
 from pyramid.view import view_config
 from pyramid.view import view_defaults
-from pyramid.httpexceptions import HTTPFound
+from pyramid.httpexceptions import HTTPFound, HTTPForbidden, HTTPNotFound
+from . import StaticPageNotFound
 from ..plugins.api import get_widget_aggregator_dispatcher
 from altaircms.helpers.viewhelpers import RegisterViewPredicate
 from altaircms.helpers.viewhelpers import FlashMessage
@@ -38,6 +39,22 @@ from altaircms.widget.forms import WidgetDispositionSaveForm
 class AfterInput(Exception):
     pass
 
+
+def page_editable(info, request):
+    if not info.get_access_status("important_page_create", "page_create"):
+        raise HTTPForbidden("not enough permission to edit it")
+    return True
+
+def page_viewable(info, request):
+    if not info.get_access_status("important_page_read", "page_read"):
+        raise HTTPForbidden("not enough permission to view it")
+    return True
+
+def page_editable_from_pagetype(context, pagetype):
+    return context.get_access_status_from_pagetype(pagetype, "important_page_create", "page_create")
+def page_viewable_from_pagetype(context, pagetype):
+    return context.get_access_status_from_pagetype(pagetype, "important_page_read", "page_read")
+
 ##
 ## todo: CRUDのview整理する
 ##
@@ -63,7 +80,8 @@ class PageAddView(object):
         self.request._setup_form = forms.PageInfoSetupWithEventForm()
         raise AfterInput
 
-    @view_config(route_name="page_add_orphan", request_param="pagetype", request_method="GET", match_param="action=input", permission="page_create")
+    @view_config(route_name="page_add_orphan", request_param="pagetype", request_method="GET", match_param="action=input", permission="page_create", 
+                 custom_predicates=[page_editable])
     def input_form(self):
         set_endpoint(self.request)
         self.request._form = forms.PageForm()
@@ -100,7 +118,8 @@ class PageAddView(object):
             raise AfterInput
 
     @view_config(route_name="page_add_orphan", permission="page_create", match_param="action=confirm", request_method="POST", 
-                 renderer="altaircms:templates/page/confirm.html")
+                 renderer="altaircms:templates/page/confirm.html", 
+                 custom_predicates=[page_editable])
     def confirm(self):
         self.request.POST
         form = forms.PageForm(self.request.POST)
@@ -133,7 +152,8 @@ class PageAddView(object):
             self.request._setup_form = forms.PageInfoSetupForm(name=form.data["name"])
             raise AfterInput
 
-    @view_config(route_name="page_add_orphan", permission="page_create", request_method="POST", match_param="action=create")
+    @view_config(route_name="page_add_orphan", permission="page_create", request_method="POST", match_param="action=create", 
+                 custom_predicates=[page_editable])
     def create_page(self):
         try:
             logging.debug('create_page')
@@ -152,8 +172,6 @@ class PageAddView(object):
             logger.exception(str(e))
             raise
 
-
- 
 
 @view_defaults(permission="page_create", decorator=with_bootstrap)
 class PageDuplicateView(object):
@@ -181,11 +199,15 @@ class PageDuplicateView(object):
     @view_config(route_name="page_duplicate", request_method="GET", renderer="altaircms:templates/page/duplicate_confirm.html")
     def duplicate_confirm(self):
         page = get_or_404(self.request.allowable(Page), Page.id==self.request.matchdict["id"])
+        if not page_editable_from_pagetype(self.context, page.pagetype):
+            raise HTTPForbidden("not enough permission to edit it")
         return {"page": page}
         
     @view_config(route_name="page_duplicate", request_method="POST")
     def duplicate(self):
         page = get_or_404(self.request.allowable(Page), Page.id==self.request.matchdict["id"])
+        if not page_editable_from_pagetype(self.context, page.pagetype):
+            raise HTTPForbidden("not enough permission to edit it")
         self.context.clone_page(page)
         FlashMessage.success(u"ページをコピーしました", request=self.request)
         return HTTPFound(self.request.GET.get("endpoint") or "/")
@@ -199,11 +221,15 @@ class PageDeleteView(object):
     @view_config(renderer="altaircms:templates/page/delete_confirm.html", request_method="GET")
     def delete_confirm(self):
         page = get_or_404(self.request.allowable(Page), Page.id==self.request.matchdict["id"])
+        if not page_editable_from_pagetype(self.context, page.pagetype):
+            raise HTTPForbidden("not enough permission to edit it")
         return {"page": page}
 
     @view_config(request_method="POST")
     def delete(self):
         page = get_or_404(self.request.allowable(Page), Page.id==self.request.matchdict["id"])
+        if not page_editable_from_pagetype(self.context, page.pagetype):
+            raise HTTPForbidden("not enough permission to edit it")
         self.context.delete_page(page)
 
         ## flash messsage
@@ -220,11 +246,15 @@ class PageSetDeleteView(object):
     @view_config(renderer="altaircms:templates/pagesets/delete_confirm.html", request_method="GET")
     def delete_confirm(self):
         pageset = get_or_404(self.request.allowable(PageSet), PageSet.id==self.request.matchdict["pageset_id"])
+        if not page_editable_from_pagetype(self.context, pageset.pagetype):
+            raise HTTPForbidden("not enough permission to edit it")
         return {"pageset": pageset, "myhelpers": myhelpers}
 
     @view_config(request_method="POST")
     def delete(self):
         pageset = get_or_404(self.request.allowable(PageSet), PageSet.id==self.request.matchdict["pageset_id"])
+        if not page_editable_from_pagetype(self.context, pageset.pagetype):
+            raise HTTPForbidden("not enough permission to edit it")
         try:
             self.context.delete_pageset(pageset)
             ## Integritty errorをキャッチしたいので
@@ -260,6 +290,8 @@ class PageUpdateView(object):
     @view_config(request_method="GET")
     def input(self):
         page = get_or_404(self.request.allowable(Page), Page.id==self.request.matchdict["id"])
+        if not page_editable_from_pagetype(self.context, page.pagetype):
+            raise HTTPForbidden("not enough permission to edit it")
         params = page.to_dict()
         form = forms.PageUpdateForm(**params)
         return self._input_page(page, form)
@@ -269,6 +301,8 @@ class PageUpdateView(object):
     def update_confirm(self):
         form = forms.PageUpdateForm(self.request.POST)
         page = get_or_404(self.request.allowable(Page), Page.id==self.request.matchdict["id"])
+        if not page_editable_from_pagetype(self.context, page.pagetype):
+            raise HTTPForbidden("not enough permission to edit it")
         if form.validate():
             return dict( page=page, params=self.request.POST.items())
         else:
@@ -277,6 +311,8 @@ class PageUpdateView(object):
     @view_config(request_method="POST", custom_predicates=[RegisterViewPredicate.execute])
     def update(self):
         page = get_or_404(self.request.allowable(Page), Page.id==self.request.matchdict["id"])
+        if not page_editable_from_pagetype(self.context, page.pagetype):
+            raise HTTPForbidden("not enough permission to edit it")
         form = forms.PageUpdateForm(self.request.POST)
         if form.validate():
             page = self.context.update_page(page, form)
@@ -301,6 +337,8 @@ class PagePartialUpdateAPIView(object):
         page = self.request.allowable(Page).filter_by(id=self.request.matchdict["id"]).first()
         if page is None:
             return {"status": False, "message": "page is not found", "data": {"layout_id": layout_id}}
+        if not page_editable_from_pagetype(self.context, page.pagetype):
+            raise HTTPForbidden("not enough permission to edit it")
         page.layout = layout
         DBSession.add(page)
         return {"status": True, "data": {"layout_id": layout_id}}
@@ -323,8 +361,8 @@ class ListView(object):
         self.context = context
         self.request = request
     
-    @view_config(renderer="altaircms:templates/pagesets/static_pageset_list.html", 
-                 custom_predicates=(dispatch_with_pagetype(lambda pagetype: pagetype.page_role == "static"), ))
+    @view_config(match_param="pagetype=static", renderer="altaircms:templates/pagesets/static_pageset_list.html", 
+                 custom_predicates=[page_viewable])
     def static_page_list(self):
         pagetype = get_pagetype(self.request)
         static_directory = get_static_page_utility(self.request)
@@ -333,11 +371,12 @@ class ListView(object):
                 "pages": pages, 
                 "pagetype": pagetype}
 
-    @view_config(renderer="altaircms:templates/pagesets/event_pageset_list.html", 
-                 custom_predicates=(dispatch_with_pagetype(lambda pagetype: pagetype.page_role == "event_detail"), ))
+
+    @view_config(match_param="pagetype=event_detail", renderer="altaircms:templates/pagesets/event_pageset_list.html", 
+                 custom_predicates=[page_viewable])
     def event_bound_page_list(self):
         """ event詳細ページと結びついているpage """
-        pagetype = get_pagetype(self.request)
+        pagetype = self.context.pagetype
         qs = self.request.allowable(PageSet).filter(PageSet.pagetype_id==pagetype.id, 
                                                     PageSet.event_id!=None)
         params = dict(self.request.GET)
@@ -354,7 +393,8 @@ class ListView(object):
         pages = h.paginate(self.request, qs, item_count=qs.count(), items_per_page=50)
         return {"pages": pages, "pagetype": pagetype, "search_form": search_form}
 
-    @view_config(renderer="altaircms:templates/pagesets/other_pageset_list.html")
+    @view_config(renderer="altaircms:templates/pagesets/other_pageset_list.html", 
+                 custom_predicates=[page_viewable])
     def other_page_list(self):
         """event詳細ページとは結びついていないページ(e.g. トップ、カテゴリトップ) """
         pagetype = get_pagetype(self.request)
@@ -373,54 +413,54 @@ class ListView(object):
         pages = h.paginate(self.request, qs, item_count=qs.count(), items_per_page=50)
         return {"pages": pages, "pagetype": pagetype, "search_form": search_form}
 
-
-def with_pageset_predicate(kind): #don't support static page
-    def decorate(info, request):
-        if not hasattr(request, "_pageset"):
-            pageset_id = request.matchdict["pageset_id"]
-            request._pageset = request.allowable(PageSet).filter_by(id=pageset_id).first()
-        pageset = request._pageset
-        if pageset is None:
-            return False
-        if pageset.event_id == None:
-            return kind == "other"
-        else:
-            return kind == "event"
-    return decorate
+from altaircms import security
+class EventPageFound(Exception, security.RootFactory):
+    def __init__(self, pageset):
+        self.pageset = pageset
 
 @view_defaults(permission="page_read", route_name="pageset_detail", decorator=with_bootstrap, request_method="GET")
 class PageSetDetailView(object):
-    def __init__(self, request):
+    def __init__(self, context, request):
+        self.context = context
         self.request = request
 
     def static_page_detail(self):
         pass
 
-    @view_config(renderer="altaircms:templates/pagesets/event_page_detail.html", 
-                 custom_predicates=(with_pageset_predicate("event"),))
+    @view_config(renderer="altaircms:templates/pagesets/event_page_detail.html", context=EventPageFound)
     def event_bound_page_detail(self):
         """ event詳細ページと結びついているpage """
-        pageset = self.request._pageset ## predicate
+        pageset = DBSession.merge(self.context.pageset) ## predicate
+        pagetype = pageset.pagetype
+        if not page_viewable_from_pagetype(self.request.context, pagetype):
+            raise HTTPForbidden("not enough permission to view it")
         return {"pageset":pageset, 
                 "myhelpers": myhelpers}
 
-
-    @view_config(renderer="altaircms:templates/pagesets/other_page_detail.html", 
-                 custom_predicates=(with_pageset_predicate("other"),))
+    @view_config(renderer="altaircms:templates/pagesets/other_page_detail.html")
     def other_page_detail(self):
         """event詳細ページとは結びついていないページ(e.g. トップ、カテゴリトップ) """
+        self.request._pageset = get_or_404(self.request.allowable(PageSet), PageSet.id==self.request.matchdict["pageset_id"])
         pageset = self.request._pageset ## predicate
+        pagetype = pageset.pagetype
+        if pagetype is None:
+            raise HTTPNotFound("pagetype is not found")
+        if pagetype.is_event_detail:
+            raise EventPageFound(pageset)
+        if not page_viewable_from_pagetype(self.context, pagetype):
+            raise HTTPForbidden("not enough permission to view it")
         return {"pageset":pageset, 
                 "myhelpers": myhelpers}
-
 
 
 @view_config(route_name="page_detail", renderer='altaircms:templates/page/view.html', permission='authenticated', 
              decorator=with_fanstatic_jqueries.merge(with_bootstrap))
-def page_detail(request):
+def page_detail(context, request):
     """ page詳細ページ
     """
     page = get_or_404(request.allowable(Page), Page.id==request.matchdict["page_id"])
+    if not page_viewable_from_pagetype(context, page.pagetype):
+        raise HTTPForbidden("not enough permission to view it")
     return {"page": page, "myhelpers": myhelpers}
 
 
@@ -433,6 +473,8 @@ def page_edit(context, request):
     """pageの中をwidgetを利用して変更する
     """
     page = request._page = getattr(request, "_page", None) or get_or_404(request.allowable(Page), Page.id==request.matchdict["page_id"])
+    if not page_editable_from_pagetype(context, page.pagetype):
+        raise HTTPForbidden("not enough permission to edit it")
     try:
         page.valid_layout()
     except ValueError, e:
